+++ docs/CHANGELOG.md
@@ -86,6 +86,9 @@ Version 1.3.0 (Bakeapple): Not yet released
 - Add `Hold Open` property to movers
 - Update `af_game_info_ext` packet extension to include bot, human, browser, and client counts
 - Display client types and counts when an AF v1.3 server is selected in the server list
+- Make team balancing distribute bots evenly across teams and exclude spectators and browsers
+- Exclude spectators and browsers from bot decommission player count calculations
+- Exclude spectators and browsers when counting players for team selection for a newly joining player
 
 [@is-this-c](https://github.com/is-this-c)
 - Use 64-bit integers for time deltas